Modify: feed mock data logic 변경
