+++ src/pages/seokyoung/Main/comment/Comment.js
@@ -1,11 +1,11 @@
 import React from "react";
 
-function Comment({ comment, deleteFunction }) {
+function Comment({ comment }) {
   return (
     <li className="commentPost">
       <span>{comment.userName}</span>
       <div>{comment.content}</div>
-      <i onClick={deleteFunction} className="fa-solid fa-xmark deleteBtn" />
+      <i className="fa-solid fa-xmark deleteBtn" />
     </li>
   );
 }
