+++ src/pages/seokyoung/Main/mainFeed/MainFeed.js
@@ -2,59 +2,22 @@ import { useEffect, useState } from "react";
 import FeedPost from "./feedPost/FeedPost";
 
 function MainFeed() {
-  const [comments, setComments] = useState([]);
-  const [inputValue, setInputValue] = useState("");
-  const [keyNum, setKeyNum] = useState(0);
-  const [feed, setFeed] = useState([
-    { feedName: "syoung_h", img: "images/seokyoung/ddol.JPG" },
-  ]);
-
-  useEffect(() => {
-    fetch("http://localhost:3000/data/common/commentData.json")
-      .then(res => res.json())
-      .then(res => {
-        setComments(res);
-        setKeyNum(res.length + 1);
-      });
-  }, []);
+  const [feed, setFeed] = useState([{ feedName: "", img: "" }]);
 
   useEffect(() => {
     fetch("http://localhost:3000/data/seokyoung/feedData.json")
       .then(res => res.json())
       .then(res => {
-        console.log(res);
         setFeed(res);
       });
   }, []);
 
-  function inputChange(e) {
-    setInputValue(e.target.value);
-  }
-
-  function addComment(e) {
-    e.preventDefault();
-    const copy = inputValue.slice().trim();
-    if (copy !== "") {
-      setComments([
-        ...comments,
-        { id: keyNum, userName: "석영", content: inputValue },
-      ]);
-    }
-    setKeyNum(prev => prev + 1);
-    setInputValue("");
-  }
-
-  return (
+  const mapMain = (
     <div className="feedMain">
-      <FeedPost
-        feed={feed}
-        comments={comments}
-        inputValue={inputValue}
-        inputChange={inputChange}
-        addComment={addComment}
-      />
+      <FeedPost feed={feed} />
     </div>
   );
+  return mapMain;
 }
 
 export default MainFeed;
